# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Visa Explorer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/search:
    get:
      operationId: searchUsers
      summary: Search users by name, with optional filters
      security:
        - session: []
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: sex
          in: query
          required: false
          schema:
            type: string
        - name: location
          in: query
          required: false
          schema:
            type: string
        - name: minAge
          in: query
          required: false
          schema:
            type: integer
        - name: maxAge
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Matching users with friendship status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSearchResult"
        "401":
          description: Unauthorized
  /users/{userId}:
    get:
      operationId: getPublicUserProfile
      summary: Get a user's public profile
      security:
        - session: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public profile with friendship status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicUserProfile"
        "401":
          description: Unauthorized
        "404":
          description: User not found
...
  /users/{userId}/friends:
    get:
      operationId: listUserFriends
      summary: Get a user's public friends list
      security:
        - session: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of user's friends
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FriendProfile"
        "404":
          description: User not found
  /friends:
    get:
      operationId: listFriends
      summary: List accepted friends
      security:
        - session: []
      responses:
        "200":
          description: List of friends
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FriendProfile"
        "401":
          description: Unauthorized
  /friends/requests:
    get:
      operationId: listFriendRequests
      summary: List incoming pending friend requests
      security:
        - session: []
      responses:
        "200":
          description: List of pending requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FriendRequest"
        "401":
          description: Unauthorized
  /friends/request/{userId}:
    post:
      operationId: sendFriendRequest
      summary: Send or accept a friend request
      security:
        - session: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Request status
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
        "401":
          description: Unauthorized
  /friends/accept/{requesterId}:
    post:
      operationId: acceptFriendRequest
      summary: Accept a friend request
      security:
        - session: []
      parameters:
        - name: requesterId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
        "401":
          description: Unauthorized
  /friends/decline/{requesterId}:
    post:
      operationId: declineFriendRequest
      summary: Decline a friend request
      security:
        - session: []
      parameters:
        - name: requesterId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
        "401":
          description: Unauthorized
  /friends/{userId}:
    delete:
      operationId: removeFriend
      summary: Unfriend a user
      security:
        - session: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    UserSearchResult:
      type: object
      required: [id]
      properties:
        id:
          type: string
        username:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        profileImageUrl:
          type: string
          nullable: true
        homeCountry:
          type: string
          nullable: true
        age:
          type: integer
          nullable: true
        sex:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        friendshipStatus:
          type: string
          nullable: true
        iRequested:
          type: boolean
          nullable: true
    PublicUserProfile:
      type: object
      required: [id]
      properties:
        id:
          type: string
        username:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        profileImageUrl:
          type: string
          nullable: true
        homeCountry:
          type: string
          nullable: true
        age:
          type: integer
          nullable: true
        sex:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        visitedCount:
          type: integer
          nullable: true
        wantToVisitCount:
          type: integer
          nullable: true
        friendshipStatus:
          type: string
          nullable: true
        iRequested:
          type: boolean
          nullable: true
    FriendProfile:
      type: object
      required: [id]
      properties:
        id:
          type: string
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        profileImageUrl:
          type: string
          nullable: true
        homeCountry:
          type: string
          nullable: true
        friendshipSince:
          type: string
          nullable: true
    FriendRequest:
      type: object
      required: [id]
      properties:
        id:
          type: string
        requesterId:
          type: string
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        profileImageUrl:
          type: string
          nullable: true
        homeCountry:
          type: string
          nullable: true
        createdAt:
          type: string
...